% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_per_hr.R
\name{at_bats_per_homerun}
\alias{at_bats_per_homerun}
\title{Calculate At-Bats Per Home Run}
\usage{
at_bats_per_homerun(at_bats, homeruns)
}
\arguments{
\item{at_bats}{Numeric vector, the total number of at-bats for each player.}

\item{homeruns}{Numeric vector, the total number of home runs for each player.}
}
\value{
Numeric vector, the average number of at-bats per home run for each player,
        rounded to one decimal place.
}
\description{
This function calculates the average number of at-bats required for a player
to hit one home run for each pair of homeruns and at_bats in the vectors provided.
It checks that all entries in the vectors are non-negative numerics and that
all homerun entries are positive.
}
\examples{
at_bats_per_homerun(c(8, 400), c(1, 20))
at_bats_per_homerun(c(-300, 400), c(10, 20)) # Error: 'at_bats' must be non-negative.
at_bats_per_homerun(c(300, 400), c(-5, 20))  # Error: 'homeruns' must be positive.
}
