% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_innings.R
\name{convert_innings}
\alias{convert_innings}
\title{Convert Innings Pitched from Recorded to Decimal Format}
\usage{
convert_innings(innings_pitched)
}
\arguments{
\item{innings_pitched}{Numeric vector, innings pitched in recorded format.}
}
\value{
Numeric or numeric vector, innings pitched in decimal format.
}
\description{
This function converts innings pitched from the traditional recording format (where fractions
of innings are recorded as 0.1 and 0.2 for one and two outs, respectively) to a decimal format
where fractions represent parts of three outs in an inning (e.g., 0.33 and 0.67). Inputs that
are not 0.1 or 0.2 in the fractional part are considered to have zero outs in that fraction.
}
\examples{
convert_innings(5.1)  # Expected output: 5.33
convert_innings(6.2)  # Expected output: 6.67
convert_innings(7)  # Expected output: 7
convert_innings(c(5.1, 6.2, 7))  # Expected output: c(5.33, 6.67, 7.00)
convert_innings("5.1")  # Error: 'innings_pitched' must be numeric.
}
