% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era.R
\name{era}
\alias{era}
\title{Calculate Earned Run Average (ERA)}
\usage{
era(earned_runs, innings_pitched)
}
\arguments{
\item{earned_runs}{Numeric vector, the number of earned runs allowed by the pitcher.}

\item{innings_pitched}{Numeric vector, the innings pitched in recorded format.}
}
\value{
Numeric vector, the ERA of the pitcher.
}
\description{
This function calculates the Earned Run Average (ERA), which is the mean number of earned runs
a pitcher would give up over nine innings, based on the earned runs allowed and the innings pitched.
Innings are adjusted from the traditional recorded format to a decimal format for accurate calculation.
}
\examples{
era(5, 9)
era(c(5, 3), c(9, 7.2))
}
