% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h9.R
\name{hits_per9}
\alias{hits_per9}
\title{Calculate Hits Per Nine Innings Pitched}
\usage{
hits_per9(hits, innings_pitched)
}
\arguments{
\item{hits}{Numeric vector, the number of hits allowed by the pitcher.}

\item{innings_pitched}{Numeric vector, the innings pitched in recorded format.}
}
\value{
Numeric vector, the number of hits per nine innings, rounded to one decimal place.
}
\description{
This function calculates the number of hits allowed per nine innings pitched. It uses the
`convert_innings` function to adjust the innings pitched from the traditional recording format
to a decimal format for accurate calculation.
}
\examples{
hits_per9(8, 9)
hits_per9(c(8, 5), c(9 ,7.2))
}
