% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr9.R
\name{homeruns_per9}
\alias{homeruns_per9}
\title{Calculate Home Runs Per Nine Innings Pitched}
\usage{
homeruns_per9(home_runs, innings_pitched)
}
\arguments{
\item{home_runs}{Numeric vector, the number of home runs allowed by the pitcher.}

\item{innings_pitched}{Numeric vector, the innings pitched in recorded format.}
}
\value{
Numeric vector, the number of home runs per nine innings, rounded to one decimal place.
}
\description{
This function calculates the number of home runs allowed per nine innings pitched. It uses the
`convert_innings` function to adjust the innings pitched from the traditional recording format
to a decimal format before performing the calculation.
}
\examples{
homeruns_per9(3, 9)
homeruns_per9(c(3, 4), c(9, 7.2))
}
