% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolated_power.R
\name{isolated_power}
\alias{isolated_power}
\title{Calculate Isolated Power}
\usage{
isolated_power(
  at_bats = NULL,
  hits = NULL,
  total_bases = NULL,
  batting_average = NULL,
  slugging_percentage = NULL
)
}
\arguments{
\item{at_bats}{Numeric vector, the number of at bats.}

\item{hits}{Numeric vector, the number of hits.}

\item{total_bases}{Numeric vector, the number of total bases achieved.}

\item{batting_average}{Numeric vector, the batting average (optional).}

\item{slugging_percentage}{Numeric vector, the slugging percentage (optional).}
}
\value{
Numeric vector, the isolated power value.
}
\description{
This function calculates isolated power (ISO), a measure of a batter's raw power, defined as
the number of extra bases per at bat. The function can calculate batting average and slugging
percentage if not provided, assuming that the necessary other statistics (at_bats, hits, and
total_bases) are available.
}
\examples{
isolated_power(at_bats = c(500, 450), hits = c(150, 135), total_bases = c(250, 230))
isolated_power(at_bats = 500, total_bases = 250, batting_average = 0.300)
}
