% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slg.R
\name{slugging_percentage}
\alias{slugging_percentage}
\title{Calculate Slugging Percentage}
\usage{
slugging_percentage(total_bases, at_bats)
}
\arguments{
\item{total_bases}{Numeric vector, the total number of bases achieved.}

\item{at_bats}{Numeric vector, the number of at bats.}
}
\value{
Numeric vector, the slugging percentage rounded to three decimal places. If the number of
        at bats is zero, the function returns NA to prevent division by zero.
}
\description{
This function calculates the slugging percentage, a measure of the total bases a batter achieves
per at bat. It is calculated as the total number of bases divided by at bats.
}
\examples{
slugging_percentage(120, 300)
slugging_percentage(c(120, 150), c(300, 350))
}
