% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strikeout_percentage.R
\name{strikeout_percentage}
\alias{strikeout_percentage}
\title{Calculate Strikeout Percentage}
\usage{
strikeout_percentage(strikeouts, at_bats)
}
\arguments{
\item{strikeouts}{Numeric vector, the number of strikeouts.}

\item{at_bats}{Numeric vector, the number of at-bats.}
}
\value{
Numeric vector, the strikeout percentage rounded to one decimal place.
        If the number of at-bats is zero, the function returns NA to prevent division by zero.
}
\description{
This function computes the strikeout percentage, which is the proportion of at-bats that result
in strikeouts, expressed as a percentage. It is calculated as the number of strikeouts divided
by the number of at-bats, multiplied by 100.
}
\examples{
strikeout_percentage(50, 200)
strikeout_percentage(c(50, 30), c(200, 100))
}
