% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so9.R
\name{strikeouts_per9}
\alias{strikeouts_per9}
\title{Calculate Strikeouts Per Nine Innings Pitched}
\usage{
strikeouts_per9(strikeouts, innings_pitched)
}
\arguments{
\item{strikeouts}{Numeric vector, the number of strikeouts.}

\item{innings_pitched}{Numeric vector, the innings pitched in traditional format.}
}
\value{
Numeric vector, the strikeouts per nine innings, rounded to one decimal place.
}
\description{
This function calculates the number of strikeouts a pitcher averages per nine innings pitched.
It first adjusts the innings pitched from a traditional recording format to a decimal format
for precise calculation.
}
\examples{
strikeouts_per9(50, 45.1)
strikeouts_per9(c(50, 75), c(45.1, 50.2))
}
