% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_percentage.R
\name{walk_percentage}
\alias{walk_percentage}
\title{Calculate Walk Percentage}
\usage{
walk_percentage(walks, at_bats)
}
\arguments{
\item{walks}{Numeric vector, the number of walks.}

\item{at_bats}{Numeric vector, the number of at-bats.}
}
\value{
Numeric vector, the walk percentage rounded to three decimal places.
        If the number of at-bats is zero, the function returns NA to prevent division by zero.
}
\description{
This function computes the walk percentage, which is the proportion of at-bats that result
in walks. It is calculated as the number of walks divided by the number of at-bats.
}
\examples{
walk_percentage(50, 200)
walk_percentage(c(50, 30), c(200, 150))
}
