% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb9.R
\name{walks_per9}
\alias{walks_per9}
\title{Calculate Walks Per Nine Innings Pitched}
\usage{
walks_per9(walks, innings_pitched)
}
\arguments{
\item{walks}{Numeric vector, the number of walks allowed by the pitcher.}

\item{innings_pitched}{Numeric vector, the innings pitched in recorded format.}
}
\value{
Numeric vector, the number of walks per nine innings, rounded to one decimal place.
}
\description{
This function calculates the number of walks per nine innings pitched.
It uses the `convert_innings` function to adjust the innings pitched from the traditional format
to a decimal format before performing the calculation.
}
\examples{
walks_per9(7, 30)
walks_per9(c(7, 35), c(40, 160))
}
